# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitzwilliam Museum Collection Database API",
        "description": "Api for the Fitzwilliam Museum Collection Database",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/publications": {
            "get": {
                "tags": [
                    "Publications"
                ],
                "summary": "Retrieve publications used in the database",
                "description": "A list of publications used in the database, with pagination.",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "Darmstadt"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of the page response",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "20"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request completed successfully."
                    },
                    "400": {
                        "description": "The request cannot be processed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/publications/{publication}": {
            "get": {
                "tags": [
                    "Publications"
                ],
                "summary": "Retrieve a publication",
                "description": "An publications's details.",
                "parameters": [
                    {
                        "name": "publication",
                        "in": "path",
                        "description": "Query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "publication-112039"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request completed successfully."
                    },
                    "400": {
                        "description": "The request cannot be processed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/terminology": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve terminology used in the database",
                "description": "A list of terminology used in the database, with pagination.",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "Painting"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of the page response",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "20"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "enum",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        },
                        "example": "asc"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Fields available to return",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "admin.id,admin.created,admin.modified,name,summary_title"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Cannot be proceessed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/terminology/{terminology}": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve a term",
                "description": "A term's details.",
                "parameters": [
                    {
                        "name": "terminology",
                        "in": "path",
                        "description": "Query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "term-30009"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Fields available to return",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "admin.id,admin.created,admin.modified,name,summary_title"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Not processed"
                    }
                }
            }
        }
    }
}